send correct number of reported issues to amplitude/mixpanel

+++ app/analyze/process.go
@@ -65,7 +65,7 @@ func analyzeLogged(ctx context.Context, repoOwner, repoName, githubAccessToken s
 	}
 
 	props := map[string]interface{}{
-		"durationSeconds": time.Since(startedAt) / time.Second,
+		"durationSeconds": int(time.Since(startedAt) / time.Second),
 	}
 	if err == nil {
 		props["status"] = "ok"
